Fixing CreateConfigurationForNewContext
When I updated the function to copy the outputHandlers from the previous config, I foolishly made it copy the reference to the previous outputHandlers hashtable; not the value. So when you run a nested build, it's outputHandlers config persists after the nested build ends, and affects the outer build.

I've now corrected it to Clone() the hashtable.